Checkpoint work on a syntax extension that allows for derivation of to/from_hlist() on arbitrary structs.
In theory this would be more flexible than the existing record_struct! macro, but writing
a custom syntax extension is a nightmare, and I'm not sure it's really worth it in the end.

Shunting this off to a separate branch in case we have the time and energy to finish it.